[AIRFLOW-6471] Add pytest-instafail plugin (#7064)
We have a problem currently that if a test fails in CI we do not see the
failures immediately - only when it finishes, but when tests hang, sometimes
the details about failed tests are not shown immediately. We tried to increase
verbosity but it's not very helpful.

The pytest-instafail plugin solves the problem without increasing verbosity.

(cherry picked from commit 869ca9619c6af8103b1b6af4e1d2f0cf82d89e6f)